capture/pyproject.toml
[project]
name = "bisimulator-capture"
version = "0.1.0"
description = "Screen capture daemon for bisimulator (macOS + Windows)"
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
macos = [
    "pyobjc-core>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-Vision>=10.0",
    "Pillow>=10.0",
]
windows = [
    "mss>=9.0",
    "Pillow>=10.0",
    "rapidocr-onnxruntime>=1.3",
]
test = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/capture"]

[tool.pytest.ini_options]
testpaths = ["tests"]
